fix PHPStan Tier 3: reduce baseline from 72 to 19 errors
Fix return types, docblock mismatches, action/filter misuse, and
argument type errors across API, Base, Plugin, Theme, REST, and
trait files. Replaces void-returning array_filter callbacks with
array_walk, wraps branch-switcher action callbacks in void closures,
corrects set_repo_cache int timeouts to bool, and aligns PHPDoc
param/return annotations with actual types throughout.

+++ src/Git_Updater/Additions/Repo_List_Table.php
@@ -410,7 +410,7 @@ class Repo_List_Table extends WP_List_Table {
 			[
 				'total_items' => $total_items,  // WE have to calculate the total number of items.
 				'per_page'    => $per_page,  // WE have to determine how many items to show on a page.
-				'total_pages' => ceil( $total_items / $per_page ), // WE have to calculate the total number of pages.
+				'total_pages' => (int) ceil( $total_items / $per_page ), // WE have to calculate the total number of pages.
 			]
 		);
 	}
